// src/types/renshuuApiTypes.ts
export interface paths {
// ...
    "/schedule": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List of all user schedules */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            schedules?: components["schemas"]["Schedule"][];
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/schedule/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List of all user schedule specified by Id */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    /** @description The schedule's id */
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            schedules?: components["schemas"]["Schedule"][];
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/schedule/{id}/list": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List of terms in the user schedule specified by id */
        get: {
            parameters: {
                query?: {
                    /** @description Page of results */
                    pg?: string;
                    /** @description Group/subset from the terms */
                    group?:
                        | "all"
                        | "blocked"
                        | "studied"
                        | "notyetstudied"
                        | "cannot_study"
                        | "review_today"
                        | "mastery_1"
                        | "mastery_2"
                        | "mastery_3"
                        | "mastery_4"
                        | "mastery_5"
                        | "mastery_6"
                        | "mastery_7"
                        | "mastery_8"
                        | "mastery_9";
                };
                header?: never;
                path: {
                    /** @description The schedule's id */
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            schedules?: components["schemas"]["Schedule"][];
                            contents?: {
                                /** @example 1 */
                                result_count?: number;
                                /** @example 2 */
                                total_pg?: number;
                                /** @example 50 */
                                per_pg?: number;
                                /** @example 1 */
                                pg?: number;
                                /** @example all */
                                group?: string;
                                terms?: (
                                    | components["schemas"]["Word"]
                                    | components["schemas"]["Kanji"]
                                    | components["schemas"]["Grammar"]
                                    | components["schemas"]["SimpleSentence"]
                                )[];
                            };
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}